add a review to a spot

+++ backend/db/migrations/20230219004003-create-review.js
@@ -30,7 +30,7 @@ module.exports = {
         allowNull: false
       },
       stars: {
-        type: Sequelize.DECIMAL(2, 1),
+        type: Sequelize.INTEGER,
         allowNull: false,
         validate: {
           min: 1,
